+++ src/Retorno.php
@@ -28,8 +28,8 @@ use \CnabPHP\RetornoAbstract;
 use \CnabPHP;
 
 class Retorno extends RetornoAbstract{
-	public function __construct($conteudo){    
-		parent::__construct($conteudo);
+	public function __construct($conteudo, $is_transf = false){    
+		parent::__construct($conteudo, $is_transf);
 	}
 }
 ?>
